requirements.txt
requests>=2.20.1
bs4>=0.0.1
Pillow>=5.3.0
google-cloud-vision>=2.0.0
lxml>=4.2.1
numpy>=1.15.4
pyzmq>=17.1.2
opencv-python>=4.0.0.21
Cython>=0.29.1
hgtk>=0.1.3
faiss-cpu>=1.7.0
tensorflow==2.5.0
tensorflow-hub==0.12.0
tensorflow-text==2.5.0